Data type fix

+++ script/lib/Calculates/Calculates.h
@@ -4,19 +4,19 @@ class Calculates
 {
     private:
 
-        float wheel_distance_x_;
-        float max_rpm_;
-        float wheel_round_;
+        double wheel_distance_x_;
+        double max_rpm_;
+        double wheel_round_;
 
     public:
-        float dcmotor_rpm;
+        double dcmotor_rpm;
 
         struct vel
         {
             double linear_x;
             double angular_z;
         };
-        Calculates(int max_rpm, float wheel_diameter, float wheel_distance_x);
+        Calculates(int max_rpm, double wheel_diameter, double wheel_distance_x);
         vel get_velocities(float steer_angle, float rpm);
-        float CalculateRpm(float linear_x);
+        double CalculateRpm(double linear_x);
 };
